requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
ciso8601==2.2.0
colorama==0.4.4
colorlog==6.6.0
frozenlist==1.3.0
hikari==2.0.0.dev108
hikari-lightbulb==2.2.1
idna==3.3
multidict==6.0.2
nextcord==2.0.0a10
Pillow==9.1.0
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
PyYAML==6.0
requests==2.27.1
tinydb==4.7.0
tinydb-smartcache==2.0.0
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
uvloop>=0.16; sys_platform != "win32"
